チャートの共通部分をmixinに切り出し
