+++ src/components/SearchResultUserRanking.vue
@@ -1,5 +1,7 @@
 <script>
 import { HorizontalBar } from 'vue-chartjs'
+import HorizontalBarDefault from '@/mixins/horizontalBarDefault.vue'
+
 export default {
   name: 'search-result-title-ranking',
   props: {
@@ -8,45 +10,7 @@ export default {
     }
   },
   extends: HorizontalBar,
-  data: function () {
-    return {
-      maxDisplayNum: 15,
-      dataCollection: {
-        labels: [],
-        datasets: [
-          {
-            label: 'いいね数',
-            backgroundColor: '#2E827F',
-            data: []
-          }
-        ]
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        scales: {
-          xAxes: [
-            {
-              ticks: {
-                min: 0,
-                fontSize: 14
-              }
-            }
-          ],
-          yAxes: [
-            {
-              display: true,
-              barPercentage: 0.7,
-              categoryPercentage: 0.7,
-              ticks: {
-                fontSize: 14
-              }
-            }
-          ]
-        }
-      }
-    }
-  },
+  mixins: [HorizontalBarDefault],
   methods: {
     computeUserRanking (array) {
       let userRanking = []
